+++ src/main/java/net/Gabou/projectatmosphere/manager/ForecastGenerator.java
@@ -657,9 +657,11 @@ public class ForecastGenerator {
 
     }
 
-    public static BiomeForecast getForecast(BiomeInstanceKey key) {
+
+    static BiomeForecast getForecast(BiomeInstanceKey key) {
         return FORECAST_MAP.get(key);
     }
+
 
 
     static float getHumidityValue(BiomeInstanceKey key, long tick) {