+++ src/main/java/net/Gabou/projectatmosphere/manager/ForecastOrchestrator.java
@@ -22,6 +22,7 @@ import net.Gabou.projectatmosphere.modules.wind.WindForecastPart;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.phys.Vec3;
 
 import java.util.*;
@@ -48,8 +49,7 @@ public class ForecastOrchestrator {
             try {
                 ForecastGenerator.generateForecastForSavedRegion(level);
                 return true;
-            }
-            catch (Exception e) {
+            } catch (Exception e) {
                 ProjectAtmosphere.LOGGER.error("[Atmosphere] Failed to load saved forecast data. Regenerating from spawn...", e);
 
                 ForecastDataStorage.clearAll(level);
@@ -123,13 +123,27 @@ public class ForecastOrchestrator {
     /**
      * Called when `/atmo regen` is used
      */
-    public static void clearAndRegenerate(ServerLevel level, Set<BlockPos> centers) {
+    public static void clearAndRegenerate(ServerLevel level) {
         ForecastGenerator.clearForecasts();
         clearActiveBiomeKeys();
         ForecastDataStorage.playerData.clear();
-
+        List<ServerPlayer> players = AsyncAtmosphereService.callOnMainThread(level::players);
+        Set<BlockPos> centers = new HashSet<>();
+        for (Player player : players) {
+            BlockPos center = player.blockPosition();
+            boolean tooClose = false;
+            ForecastDataStorage.playerData.put(player.getUUID(), center);
+            for (BlockPos existingCenter : centers) {
+                if (existingCenter.distManhattan(center) < MIN_DISTANCE_BETWEEN_CENTERS) {
+                    tooClose = true;
+                    break;
+                }
+            }
+            if (!tooClose) {
+                centers.add(center);
+            }
+        }
         for (BlockPos center : centers) {
-            ForecastDataStorage.playerData.put(UUID.randomUUID(), center);
             ForecastGenerator.generateForecastForRegion(center, level);
         }
 
@@ -218,7 +232,7 @@ public class ForecastOrchestrator {
 
     public static void tick(ServerLevel level) {
         GlassDamageManager.tick(level);
-        if(sandStormLoaded)
+        if (sandStormLoaded)
             ForecastGenerator.tickSandstormScheduler(level);
 
         long now = level.getGameTime();